+++ medicom_tools/src/app/viewapp.rs
@@ -219,7 +219,7 @@ impl DicomFileImageLoader {
     fn to_image(imgvol: &ImageVolume, axis: &VolAxis, slice_index: usize) -> ColorImage {
         let win = imgvol
             .minmax_winlevel()
-            .with_out(f64::from(u8::MIN), f64::from(u8::MAX));
+            .with_out(u8::MIN as f32, u8::MAX as f32);
 
         let axis_dims = imgvol.axis_dims(axis);
